+++ windows/flutter/generated_plugins.cmake
@@ -6,8 +6,10 @@ list(APPEND FLUTTER_PLUGIN_LIST
   audioplayers_windows
   flutter_tts
   permission_handler_windows
+  screen_retriever_windows
   url_launcher_windows
   volume_controller
+  window_manager
 )
 
 list(APPEND FLUTTER_FFI_PLUGIN_LIST